requirements.txt
pyautogui
pynput
pywinauto
pillow
numpy